+++ draft-packages/stories/GuidanceBlock.stories.tsx
@@ -1,8 +1,9 @@
 import * as colorTokens from "@kaizen/design-tokens/tokens/color.json"
 import * as React from "react"
 
-import { GuidanceBlock } from "@kaizen/component-library/draft"
+import { GuidanceBlock } from "@kaizen/draft-guidance-block"
 const bannerImg = require("@kaizen/component-library/stories/illustrations/team.png")
+const meetingImg = require("@kaizen/component-library/stories/illustrations/meeting-voices.png")
 
 export default {
   title: "GuidanceBlock (React)",
@@ -17,15 +18,65 @@ const Default = () => (
         description:
           "Mussum Ipsum, cacilds vidis litro abertis. Suco de cevadiss, é um leite divinis, qui tem lupuliz, matis, aguis e fermentis. Mé faiz elementum girarzis, nisi eros vermeio.",
       }}
-      button={{
-        label: "Action",
-        onClick: () => {
-          alert("tada: 🎉")
+      actions={{
+        primary: {
+          label: "Action",
+          onClick: () => {
+            alert("tada: 🎉")
+          },
+        },
+        dismiss: {
+          onClick: () => alert("tada: 🎉"),
         },
       }}
-      onDismiss={() => {
-        /* do nothing */
+    />
+  </div>
+)
+
+const Persistent = () => (
+  <div style={{ display: "flex", justifyContent: "center", margin: "100px" }}>
+    <GuidanceBlock
+      img={{ src: meetingImg, alt: "Meeting illustration" }}
+      text={{
+        title: "You're using the updated navigation",
+        description:
+          "This beta is to test out the new navigation, read about our thinking behind the change. And let us know if you have any feedback",
       }}
+      actions={{
+        primary: {
+          label: "Turn on for the whole org",
+          onClick: () => {
+            alert("tada: 🎉")
+          },
+        },
+      }}
+      persistent
+    />
+  </div>
+)
+
+const SecondaryAction = () => (
+  <div style={{ display: "flex", justifyContent: "center", margin: "100px" }}>
+    <GuidanceBlock
+      img={{ src: meetingImg, alt: "Meeting illustration" }}
+      text={{
+        title: "You're using the updated navigation",
+        description:
+          "This beta is to test out the new navigation, read about our thinking behind the change. And let us know if you have any feedback",
+      }}
+      actions={{
+        primary: {
+          label: "Turn on for the whole org",
+          onClick: () => {
+            alert("tada: 🎉")
+          },
+        },
+        secondary: {
+          label: "Turn off new navigation",
+          onClick: () => alert("tada: 🎉"),
+        },
+      }}
+      persistent
     />
   </div>
 )
@@ -43,4 +94,30 @@ Default.story = {
   },
 }
 
-export { Default }
+Persistent.story = {
+  name: "Persistent",
+  parameters: {
+    backgrounds: [
+      {
+        name: "Stone",
+        value: colorTokens.kz.color.stone,
+        default: true,
+      },
+    ],
+  },
+}
+
+SecondaryAction.story = {
+  name: "Secondary Action",
+  parameters: {
+    backgrounds: [
+      {
+        name: "Stone",
+        value: colorTokens.kz.color.stone,
+        default: true,
+      },
+    ],
+  },
+}
+
+export { Default, Persistent, SecondaryAction }